+++ projects/fhi-angular-components/src/lib/fhi-tree-view-navigation/fhi-tree-view-navigation-item.model.ts
@@ -1,0 +1,10 @@
+export interface FhiTreeViewNavigationItem {
+  children?: FhiTreeViewNavigationItem[];
+  href?: string;
+  id?: number | string;
+  isActive?: boolean;
+  isExpanded?: boolean;
+  hasActiveDescendant?: boolean;
+  name: string;
+  routerLink?: string;
+}