+++ projects/fhi-angular-components/src/lib/fhi-tree-view-navigation/fhi-tree-view-navigation.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FhiTreeViewNavigationComponent } from './fhi-tree-view-navigation.component';
+
+describe('FhiTreeViewNavigationComponent', () => {
+  let component: FhiTreeViewNavigationComponent;
+  let fixture: ComponentFixture<FhiTreeViewNavigationComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ FhiTreeViewNavigationComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(FhiTreeViewNavigationComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});