+++ projects/fhi-angular-components/src/lib/fhi-tree-view-navigation/fhi-tree-view-navigation.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+
+import { FhiTreeViewNavigationComponent } from './fhi-tree-view-navigation.component';
+
+@NgModule({
+  declarations: [
+    FhiTreeViewNavigationComponent
+  ],
+  imports: [
+    CommonModule,
+    RouterModule
+  ],
+  exports: [
+    FhiTreeViewNavigationComponent
+  ]
+})
+export class FhiTreeViewNavigationModule { }