+++ projects/fhi-angular-components/src/public-api.ts
@@ -9,3 +9,7 @@ export * from './lib/fhi-autosuggest/fhi-autosuggest.module';
 
 export * from './lib/fhi-multiselect/fhi-multiselect.component';
 export * from './lib/fhi-multiselect/fhi-multiselect.module';
+
+export * from './lib/fhi-tree-view-navigation/fhi-tree-view-navigation.component';
+export * from './lib/fhi-tree-view-navigation/fhi-tree-view-navigation.module';
+export * from './lib/fhi-tree-view-navigation/fhi-tree-view-navigation-item.model';