+++ projects/fhi-angular-highcharts/CHANGELOG.md
@@ -1,4 +1,12 @@
-# 0.1.0
+# 0.2.0
+
+> Mar 14, 2023
+
+* :tada: **Enhancement** Update README
+* :tada: **Enhancement** Add support for Angular 15
+* :tada: **Enhancement** Upgrade Highcharts and "Highcharts Angular" to latest
+
+## 0.1.0
 
 > Mar 07, 2023